Do bulk ValidateAndPrepare, Commit
For many cases, validation will be faster than commit. Hence, it
causes a buildup of validated blocks (due to the pipelining).
If we have a buildup, we can batch all of the blocks together
into a single request for CouchDB - affects both ValidateAndPrepare
as well as Commit methods.

Additionally this commit also makes HistoryDB and StateDB run
in parallel.

Change-Id: Ibfa9d34d1348fa28a3f948e67dfe826f843fcedb